ANDROID: sched: Disable TTWU_QUEUE by default
TTWU_QUEUE means that when waking up a task on a remote CPU with
which we don't share a cache, instead of directly enqueueing on that
CPU we instead send an IPI asking that CPU to do the enqueue. This
avoids rq lock contention across cache domains.

The downside of TTWU_QUEUE is that when waking up multiple tasks in
quick succession, the enqueue of one task is not generally completed
before we start making task-placement decisions for the next
task. That means the target CPU's load hasn't been changed, likely
resulting in unnecessarily coscheduling of tasks, which can impact
benchmark scores.

The assumption here is that the rq lock contention poses only a
minimal thread on Android devices which typically have relatively few
CPUs, so this tradeoff for better task placement makes sense.

Change-Id: I18db618885618d3c03450e552d9e92aafacd8298
Signed-off-by: Brendan Jackman <brendan.jackman@arm.com>
Signed-off-by: Chris Redpath <chris.redpath@arm.com>
(cherry picked from commit 81605b2a8a77a1fd21fd06a23c4d277164cd8256)