cpu-boost: Introduce Dynamic Schedtune Boost
Dynamic schedtune boosting occurs on-the-fly during interactions, such as touch input. This idea was conceived because of an existing battery drain issue whereby setting a permanent schedtune boost would cause boosting to occur even when the phone was idling. For example, if a user sets the /dev/stune/top-app/schedtune.boost to 10, the device would almost never idle at the lowest frequency step, which resulted in higher idle and active drain rates. The advantage of the Dynamic Schedtune Boost approach is that it allows your device to boost only when it really needs to.

The tunable can be found in /sys/module/cpu_boost/parameters/dynamic_stune_boost.

Signed-off-by: joshuous <joshuous@gmail.com>
(cherry picked from commit 490847ac07e8ce4498798e9515e73bbc5aeaacd9)